+++ src/pages/addCard/components/AddCardForm.tsx
@@ -6,36 +6,17 @@ import CardHolderNameInput from "./CardHolderNameInput";
 import CardPasswordInput from "./CardPasswordInput";
 import useCardNumberInput from "../hooks/useCardNumberInput";
 
-const CARN_NUMBER_LENGTH = 16;
-
 export default function AddCardForm() {
-  const [cardNumber, setCardNumber] = useState<string>("");
+  const {
+    cardNumber,
+    displayedCardNumber,
+    handleChange: handleCardNumberChange,
+  } = useCardNumberInput();
 
   const [cardExpiredDate, setCardExpiredDate] = useState<string>("");
   const [cardHolderName, setCardHolderName] = useState<string>("");
   const [cardSecurityCode, setCardSecurityCode] = useState<string>("");
   const [cardPassword, setCardPassword] = useState<string>("");
-
-  const { toMaskedCardNumber } = useCardNumberInput();
-
-  const displayedCardNumber = useMemo(
-    () => toMaskedCardNumber(cardNumber),
-    [cardNumber, toMaskedCardNumber]
-  );
-
-  const handleCardNumberChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    const lastInputKey = value.slice(value.length - 1);
-    console.log(value.length);
-    console.log(displayedCardNumber.length);
-    if (cardNumber.length === CARN_NUMBER_LENGTH) return;
-    if (value.length > displayedCardNumber.length) {
-      if (isNaN(Number(lastInputKey))) return;
-      setCardNumber((prev) => prev + lastInputKey);
-      return;
-    }
-    setCardNumber((prev) => prev.slice(0, cardNumber.length - 1));
-  };
 
   const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -46,14 +27,14 @@ export default function AddCardForm() {
         value={displayedCardNumber}
         onChange={handleCardNumberChange}
       />
-      <CardExpiredDateInput
+      {/* <CardExpiredDateInput
         cardExpiredDate={cardExpiredDate}
         setCardExpiredDate={setCardExpiredDate}
-      />
-      <CardHolderNameInput
+      /> */}
+      {/* <CardHolderNameInput
         cardHolderName={cardHolderName}
         setCardHolderName={setCardHolderName}
-      />
+      /> */}
       {/* <CardSecurityCodeInput cardSecurityCodeRef={cardSecurityCodeRef} />
       <CardPasswordInput
         firstLetterRef={firstLetterOfPasswordRef}
